Changed a few 'int's to 'size_t'.  Hopefully this will take care of the issue.  I don't have enough memory to test it on Lane's file though.

+++ apbs/tools/mesh/dx2uhbd.c
@@ -15,7 +15,7 @@ VEMBED(rcsid="$Id$")
 int main(int argc, char **argv) {
 
   /*** Variables ***/
-  int u, i, j, k, nx, ny, nz;
+  size_t u, i, j, k, nx, ny, nz;
   double avg;
   double hy, hx, hzed, xmin, ymin, zmin;
   Vio *sock;
